services/live/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
aiortc
numpy==1.26.3
scipy==1.12.0
python-multipart==0.0.6
structlog==24.1.0
requests==2.31.0
websockets==12.0
deepgram-sdk==3.2.0
google-generativeai==0.3.2
